requirements.txt
# ============================================================
# Multi-Agent AI Academic Project Assistant - Requirements
# ============================================================

# Core LLM & AI
google-generativeai>=0.7.0
openai>=1.40.0

# RAG & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0

# Document Processing
pypdf2>=3.0.1
PyMuPDF>=1.24.0
python-docx>=1.1.0
tavily-python>=0.3.3
python-pptx>=0.6.23

# Frontend
streamlit>=1.35.0
streamlit-extras>=0.4.0

# Backend (optional FastAPI layer)
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0

# Utilities
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
tenacity>=8.3.0
cachetools>=5.3.0
rich>=13.7.0
loguru>=0.7.0

# Presentation
Pillow>=10.3.0
matplotlib>=3.9.0
